+++ aioplus/__init__.py
@@ -17,6 +17,8 @@ from aioplus.internal.aminmax import aminmax
 from aioplus.internal.anextify import anextify
 from aioplus.internal.anth import anth
 from aioplus.internal.apairwise import apairwise
+from aioplus.internal.apostpend import apostpend
+from aioplus.internal.aprepend import aprepend
 from aioplus.internal.arace import arace
 from aioplus.internal.arange import arange
 from aioplus.internal.arepeat import arepeat
@@ -55,6 +57,8 @@ __all__: list[str] = [
     "anextify",
     "anth",
     "apairwise",
+    "apostpend",
+    "aprepend",
     "arace",
     "arange",
     "arepeat",
@@ -89,6 +93,8 @@ aminmax.__module__ = "aioplus"
 anextify.__module__ = "aioplus"
 anth.__module__ = "aioplus"
 apairwise.__module__ = "aioplus"
+apostpend.__module__ = "aioplus"
+aprepend.__module__ = "aioplus"
 arace.__module__ = "aioplus"
 arange.__module__ = "aioplus"
 arepeat.__module__ = "aioplus"